requirements.txt
fake-bpy-module-latest
pylint==3.3.6
watchdog==6.0.0